requirements.txt
Appium-Python-Client==2.11.1
async-generator==1.10
allure-pytest==2.9.45
attrs==23.1.0
behave==1.2.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
pycparser==2.21
PySocks==1.7.1
pytest==7.3.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.3
urllib3==2.0.3
webdriver-manager==3.8.6
wsproto==1.2.0